requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
asttokens==3.0.0
    # via stack-data
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6
    # via ipython
decorator==5.1.1
    # via ipython
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.1.6
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==65.4.1
    # via -r requirements.in
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.4.0
    # via -r requirements.in
executing==2.2.0
    # via stack-data
greenlet==3.1.1
    # via sqlalchemy
idna==3.10
    # via requests
ipython==8.32.0
    # via -r requirements.in
jedi==0.19.2
    # via ipython
matplotlib-inline==0.1.7
    # via ipython
numpy==2.2.2
    # via pandas
pandas==2.2.3
    # via -r requirements.in
parso==0.8.4
    # via jedi
pillow==11.1.0
    # via -r requirements.in
prompt-toolkit==3.0.50
    # via ipython
pure-eval==0.2.3
    # via stack-data
pygments==2.19.1
    # via ipython
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.1
    # via pandas
requests==2.32.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.38
    # via -r requirements.in
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via sqlalchemy
tzdata==2025.1
    # via
    #   django
    #   pandas
urllib3==2.3.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
